Move CssRuleSensor into 'eslint' package
